Stub of the emitter.
The new approach is to emit serialization code. During emitting,
as much as possible is done to be efficient during serialization.
For instance, if integer field is to be written, we can directly
emit load field, write integer; if the field has a type object,
then we have to perform check at serialization time. We cannot be
sure what actual type of that object is.