Reworked dictionary serialization.
It was required to use strict reflection instead of dynamic/Impromptu,
because of situation when a class implements multiple different
dictionary/ienumerable interfaces.

This case has been observed with IronPython's PythonDictionary. One of
the implementations of GetEnumerator is inconsistent with the others
and, instead of returning KeyValuePair, it returned only the key of an
entry.