Go back again to the version where late is default.
Motivation is that it is extremly intuitive to have all referenced objects deserialized when post hook is invoked. Also list was changed to stack, therefore hooks are invoked in a way that dependant ones will be invoked later, which is intuitive as well. Naturally, it does not handle inverted hook depedencies and circular dependecies, but what handles that.